// server/prisma/__generated__/graphql/resolvers/inputs/IntWithAggregatesFilter.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../../scalars";
import { NestedFloatFilter } from "../inputs/NestedFloatFilter";
import { NestedIntFilter } from "../inputs/NestedIntFilter";
import { NestedIntWithAggregatesFilter } from "../inputs/NestedIntWithAggregatesFilter";

@TypeGraphQL.InputType("IntWithAggregatesFilter", {})
export class IntWithAggregatesFilter {
    @TypeGraphQL.Field(_type => TypeGraphQL.Int, {
            nullable: true
        })
    equals?: number | undefined;

    @TypeGraphQL.Field(_type => [TypeGraphQL.Int], {
            nullable: true
        })
    in?: number[] | undefined;

    @TypeGraphQL.Field(_type => [TypeGraphQL.Int], {
            nullable: true
        })
    notIn?: number[] | undefined;

    @TypeGraphQL.Field(_type => TypeGraphQL.Int, {
            nullable: true
        })
    lt?: number | undefined;

    @TypeGraphQL.Field(_type => TypeGraphQL.Int, {
            nullable: true
        })
    lte?: number | undefined;

    @TypeGraphQL.Field(_type => TypeGraphQL.Int, {
            nullable: true
        })
    gt?: number | undefined;

    @TypeGraphQL.Field(_type => TypeGraphQL.Int, {
            nullable: true
        })
    gte?: number | undefined;

    @TypeGraphQL.Field(_type => NestedIntWithAggregatesFilter, {
            nullable: true
        })
    not?: NestedIntWithAggregatesFilter | undefined;

    @TypeGraphQL.Field(_type => NestedIntFilter, {
            nullable: true
        })
    _count?: NestedIntFilter | undefined;

    @TypeGraphQL.Field(_type => NestedFloatFilter, {
            nullable: true
        })
    _avg?: NestedFloatFilter | undefined;

    @TypeGraphQL.Field(_type => NestedIntFilter, {
            nullable: true
        })
    _sum?: NestedIntFilter | undefined;

    @TypeGraphQL.Field(_type => NestedIntFilter, {
            nullable: true
        })
    _min?: NestedIntFilter | undefined;

    @TypeGraphQL.Field(_type => NestedIntFilter, {
            nullable: true
        })
    _max?: NestedIntFilter | undefined;
}